Address PR comments. Revert to jest original config. Removed extra logging.
