--- jest.config.js
@@ -1,18 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  transform: {
-    "^.+\\.[jt]sx?$": "<rootDir>/jest-preprocess.js",
-  },
-  moduleNameMapper: {
-    ".+\\.(css|styl|less|sass|scss)$": `identity-obj-proxy`,
-  },
-  testPathIgnorePatterns: [`<rootDir>.*/public`, `\\.cache`, `<rootDir>.*/node_modules`],
-  transformIgnorePatterns: [`node_modules/(?!(gatsby)/)`],
-  globals: {
-    __PATH_PREFIX__: ``,
-  },
-  setupFiles: [
-    'dotenv/config',
-  ],
-}
