+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "governance",
+  "name": "snapshot",
   "version": "0.0.0-development",
   "description": "",
   "main": "Pulumi.ts",
   "scripts": {
-    "test": "jest",
+    "test": "jest --no-cache --no-watchman --runInBand",
     "build": "concurrently -c blue,green -n TYPESCRIPT,GATSBY 'tsc -p .' 'gatsby build'",
     "develop": "gatsby develop --https -H 0.0.0.0",
     "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
@@ -14,6 +14,15 @@
     "version": "tsc -p .",
     "migrate": "DOTENV_CONFIG_PATH=.env.development ts-node -r dotenv/config.js ./node_modules/node-pg-migrate/bin/node-pg-migrate -j ts -m src/migrations -d CONNECTION_STRING",
     "production": "./node_modules/node-pg-migrate/bin/node-pg-migrate -m lib/migrations -d CONNECTION_STRING up && NODE_ENV=production node lib/server.js"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "roots": [
+      "<rootDir>/src/"
+    ],
+    "moduleNameMapper": {
+      "\\.(css|less|sass|scss|gif|ttf|eot|svg)$": "<rootDir>/src/__mocks__/files.ts"
+    }
   },
   "repository": {
     "type": "git",
@@ -60,12 +69,9 @@
     "@types/numeral": "^2.0.1",
     "@types/slug": "^0.9.1",
     "@types/validator": "^13.1.3",
-    "babel-jest": "^26.6.3",
-    "babel-preset-gatsby": "^2.0.0",
     "concurrently": "^6.0.0",
     "decentraland-gatsby-deploy": "^1.9.4",
     "devcert": "1.1.0",
-    "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7",
     "postcss-assets": "^5.0.0",
