+++ src/entities/Votes/model.ts
@@ -1,8 +1,9 @@
-import { createHash } from "crypto"
-import { Model } from "decentraland-gatsby/dist/entities/Database/model"
-import { join, SQL, table } from "decentraland-gatsby/dist/entities/Database/utils"
-import { VoteAttributes } from "./types"
-import { ProposalAttributes } from "../Proposal/types"
+import { createHash } from 'crypto'
+import { Model } from 'decentraland-gatsby/dist/entities/Database/model'
+import { SQL, table, join } from 'decentraland-gatsby/dist/entities/Database/utils'
+import { VoteAttributes } from './types'
+import { ProposalAttributes } from '../Proposal/types'
+
 
 export default class VotesModel extends Model<VoteAttributes> {
   static tableName = 'scores'
@@ -35,7 +36,7 @@ export default class VotesModel extends Model<VoteAttributes> {
 
   static async safelyGetVotesFor(proposal: ProposalAttributes<any>) {
     let proposalVotes: VoteAttributes | null = await VotesModel.getVotes(proposal.id)
-    return !proposalVotes ? await VotesModel.createEmpty(proposal.id) : proposalVotes.votes
+    return proposalVotes === null || proposalVotes.votes === {} ? await VotesModel.createEmpty(proposal.id) : proposalVotes.votes
   }
 
   static newScoreFor(proposal_id: string) {
