+++ tsconfig.json
@@ -57,14 +57,7 @@
       "node_modules/@types",
       "src/@types"
     ],                       /* List of folders to include type definitions from. */
-    "baseUrl": "./",
-    "paths": {
-      "*" : ["src/@types/*"]
-    },
-    "types": [
-      "node",
-      "jest"
-    ],                          /* Type declaration files to be included in compilation. */
+    // "types": [],                           /* Type declaration files to be included in compilation. */
     // "allowSyntheticDefaultImports": true,  /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
     "esModuleInterop": true /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
     // "preserveSymlinks": true,              /* Do not resolve the real path of symlinks. */
@@ -80,8 +73,5 @@
     // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
   },
-  "include": ["src"],
-  "exclude": [
-    "test/*","**/*.test.*", "**/__tests__/*"
-  ]
+  "include": ["src"]
 }
